requirements.txt
python-telegram-bot
colorama~=0.4.6
deepdiff~=6.7.1
requests~=2.31.0
beautifulsoup4~=4.12.3
openai~=1.12.0
selenium~=4.18.1
mitmproxy
urllib3==1.26.6
python-dotenv~=1.0.1
click~=8.1.7
ipython~=8.22.1